more cleanup

+++ preprocess2/db/bubble_db.py
@@ -5,20 +5,18 @@ from preprocess2.bubble.BubbleData import BubbleData
 
 NAME = "bubbles.db"
 
-def get_connection(chr_dir):
+def get_connection(chr_dir, clear_existing=False):
     db_path = os.path.join(chr_dir, NAME)
+
+    if clear_existing and os.path.exists(db_path):
+            os.remove(db_path)
+
     conn = sqlite3.connect(db_path)
     conn.row_factory = sqlite3.Row
     return conn
 
-def _remove_if_exists(chr_dir):
-    db_path = os.path.join(chr_dir, NAME)
-    if os.path.exists(db_path):
-        os.remove(db_path)
-
 def create_bubble_tables(chr_dir):
-    _remove_if_exists(chr_dir)
-    conn = get_connection(chr_dir)
+    conn = get_connection(chr_dir, clear_existing=True)
     cur = conn.cursor()
 
     cur.execute("""
